Add abstracted log handling under utils/analyze/
Abstracted logs are fundamentally different from raw log files. This
adds a library which enables the various benchmarks' differing output
directory structures to be imported into a uniform format via one of the
import scripts.

Import scripts will have to be written for each benchmark suite, but
after that, the same existing scripts can be used to analyze the data.

---

The added .envrc file is for setting environment variables which the
project needs to function well. In this case, the util/ directory has
many python scripts and modules, meaning that without the .envrc, users
would need to manually specify the PYTHONPATH when running scripts.

This .envrc is supported by the direnv tool, which automatically loads
the .envrc file when inside a directory containing a .envrc (with
appropriate security constraints). That tool may be desirable when
working with OptSched.