Make Graph Trans application location an option
Rather than hardcoding the application before or after the heuristic,
make it a runtime option.

Applying graph transformations before the heuristic has the advantage
that the heuristic now has a simpler problem to solve, and thus is more
likely to find an optimal solution. Furthermore, the lower bound will be
tighter.

Applying graph transformations after the heuristic has the advantage
that we don't waste time running graph transformations for problems that
are already easy; if the heuristic scheduler proves the schedule
optimal, we skip graph transformations altogether.

Another option is possible: we could do the heuristic first, then if the
schedule is not known to be optimal, run graph transformations (as in
GT after heuristic), but then re-run the heuristic so that we can
potentially skip enumeration. This option is NOT implemented.